sv_chars.lua basic implementation
The idea is for sv_chars.lua to handle everything custom character from the server side.

At the start of a round, after all players are spawned, sv_chars's SetPlayerCharacters() function is called. This is where all the magic and main idea of this mode should happen.

For now it only sets the players' models and sexes from the characters table. If there are not enough characters for everyone, some will just keep their "default" models.