Custom bodygroups
This is a table containing all bodygroups you would see for the playermodel in sandbox mode, each key being the bodygroup name for its respective value.
Also "Skin" is not really a bodygroup but since it shows up in the bodygroups tab in sandbox, I decided to add it here as well.