Added and implemented name color finally
This was the hardest to do so far because I had to go through every mention of Player:GetPlayerColor(), check its context and replace with GetNameColor() accordingly. It wasn't easy bruh :sob:
There are still some places I decided not to mess with because either I don't know what it does or it has to do with the murderer's disguise mechanic and I gotta go a little deeper into its code so I don't mess anything up.