Made -byname and -byindex tags optional
Running either mwcc_char_info or mwcc_char_edit, if the first argument is a number, it'll consider it as the index for the character. If not, it'll consider it as the name. If you're having trouble with this, you can still use -byname or -byindex to explicitly state that it's the name/index.