Command for adding characters
mwcc_char_add creates a new character with some default settings and adds it to the characters table. If any other arguments are passed, it'll run mwcc_char_edit on the newly creater character with the arguments passed, making it a clever and easy way to create and set new characters.

And that's it for the character CRUD commands, now for the interface.
(wait actually i just remembered i wanted to get the bodygroups from the playermodel if you changed it with mwcc_char_edit